🚚 api: certification-attestations-for-division (third route):
move repository method to certification course src
